refactor: posts-manager 페이지를 pages폴더로 이동
